Extracted service class from job
