+++ app/jobs/xero/contact_pull_job.rb
@@ -1,32 +1,17 @@
 module Xero
   class ContactPullJob < ApplicationJob
-
     def perform
-      Contact.upsert_api_response(api_results)
+      Xero::Sync.new(api_client).perform
       # I wish we could just add the org code to each record
       # as we were saving it…
       Contact.update_all org_code: org_code
     end
 
-    def fetched_results
-      api_results.size
-    end
-
-  private
+    private
 
     # https://developer.xero.com/documentation/api/organisation
     def org_code
       @org_code ||= api_client.Organisation.first.short_code
-    end
-
-    def api_results
-      Error.wrap do
-        # https://developer.xero.com/documentation/api/contacts
-        @api_results ||= api_client.Contact.all(
-          modified_since: Contact.last_updated_at,
-          include_archived: true,
-        )
-      end
     end
 
     # https://github.com/waynerobinson/xeroizer
